% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cwebby}
\alias{theme_cwebby}
\title{Custom website plot theme}
\usage{
theme_cwebby(base_size = 12, base_family = "", vertical_y_title = FALSE)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{vertical_y_title}{boolean - do you want the y axis title to be rotated vertically?}
}
\value{
A modified theme object
}
\description{
Uses the same gray in the sidebar for panel background, and a slightly darker gray than the page body for text.
}
